pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "torch-obb"
dynamic = ["version"]
description = "PyTorch package for oriented bounding box intersection operations"
readme = "README.md"
license = "MIT"
authors = [
    {name = "torch-obb contributors"},
]
maintainers = [
    {name = "torch-obb contributors"},
]
keywords = ["pytorch", "bounding-box", "intersection", "obb", "oriented", "cuda", "warp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.20.0",
    "warp-lang>=0.10.0",
    "torch>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "numba>=0.50.0",
]
all = [
    "torch-obb[dev]",
]

[tool.setuptools]
zip-safe = false

[tool.setuptools.packages.find]
where = ["."]
include = ["torch_obb*"]

[tool.setuptools.package-data]
torch_obb = ["py.typed"]

[tool.setuptools_scm]
write_to = "torch_obb/_version.py"
